app/requirements.txt
discord.py
atproto
aiofiles
tenacity
